requirements-zero.txt
Django==3.2
pinax-templates==3.0.0
